Adds support Built-in Functions and highlighting unfinished Comments and Strings
Adds support for Built-in Functions (print, require, etc), and supports highlighting unfinished comments and strings.

Made for usage as a Roblox module, so a lot of the builtin may be specific to the Roblox API, but you can easily remove them from the list.
Also, I removed the bits about filter and options and whatnot because I didn't use that so it was faster and cleaner for me to remove it. Very quick and easy to put the if statements back in.